Safari-609.1.11.1/LayoutTests/fast/dom/html-link-element-activation-behavior-url-is-null.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<link id="target">
<script>
description("This tests html link element activation behavior when url is null");
function notNavigatedAnywhere(){
    finishJSTest();
}
var mouseEvent = new MouseEvent("click");
var target = document.getElementById("target");
shouldBeTrue("target.dispatchEvent(mouseEvent)");
setTimeout(notNavigatedAnywhere, 100);
window.jsTestIsAsync = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
