Safari-611.1.10/LayoutTests/fast/dom/html-link-element-activation-behavior-url-is-null-expected.txt

This tests html link element activation behavior when url is null

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS target.dispatchEvent(mouseEvent) is true
PASS successfullyParsed is true

TEST COMPLETE
